sentiment_analysis/wsb_tracker/requirements.txt
psaw==0.1.0
psycopg2-binary==2.8.6